Refresh the list of preloaded drawables and color state lists
Bug 3266751

The boot time is 43 compared to 44 seconds, which is within the precision of the measure.

Starting 15 activities in a sequence takes 20-24 seconds with the old version and 17-18
seconds with the new pre-load (12/13 when ran a second time, all apps being loaded).
The standard deviation is high, but it seems consistently better.

Change-Id: Ieac3cb93be03e8186979b894adda29f0549b9734