When the user touches a peeked notification, reset the fadeout timer.
We do this regardless of whether there are custom controls
mostly to keep things simple.

NB: This change also reduces the fade time to 3sec from 5,
which has always seemed too long but was necessary to make
interacting with Music possible. Now that you can keep a
notification open arbitrarily long by putting your finger on
it, a shorter timeout is acceptable.

Bug: 3456677
Change-Id: I674abfcc0fa757f838b98ed9d5d46ba785a95cfa