Fix 3G connectivity bug 3513741.
ServiceStateTracker the registerForDataConnectionDetached needed to
test for !STATE_IN_SERVICE so that the other values for state also cause
the detached notifciation.

CdmaServiceStateTracker needed to use the ServiceStateTracker lists
and not the cdma specific lists for registration, which are now deleted.
Also removed an unused import.

CdmaLteServiceStateTracker cannot use the deleted cdma specific
registration lists.


Bug: 3513741
Change-Id: I6b2da62714ed598d954f0c68f738ded0f385ade2