Improve browser scrolling when using the mouse wheel.
The way that smooth scrolling is implemented in the Browser,
repeated requests to scroll by a certain amount do not add up
to one large scroll by the cumulative amount.  This makes
the mouse wheel unusable on large pages because the Browser
will scroll at a more or less constant rate no matter how often
the wheel is turned.

The fix is to not animate scrolls when using the mouse wheel.

Change-Id: I23c05cdd2383944b8730deb225b7f3f57f1729df