Fix handling of unsol_data_state_change with PDP sharing scenario
The original logic was to go through each ApnContext to find
changes on the DataConnection if applicable. This was causing
an issue when an DC is shared by multiple ApnContexts. Only
the first ApnContext associated with the updated DC was detecting
the link properties change.

The change in this patch is to go through the changes by DC instead
of ApnContext. And make sure the update on a DC is propagated to
all associated ApnContexts.

Bug: 4744006
Change-Id: Ie52d62d1d5671005f9e74b1ddc24822c9013e3e4